# Dynamically choose lengths table type based on input size