This is a attempt at dynamically choosing the counter type based on input size as discussed in #10. 
#### Benchmark before and after:

| String Lengths | Time Before | Time After |
| --- | --- | --- |
| 20 000 chars | 2836ms | 1833ms |